Make Model Confirm into Service to reuse in other Controllers

+++ src/app/leagues/leagues.controller.js
@@ -6,10 +6,10 @@
 		.module('sportsAdmin')
 		.controller('LeaguesController', LeaguesController);
 
-	LeaguesController.$inject = ['$modal', 'initialData', 'sportsApi'];
+	LeaguesController.$inject = ['dialogsService', 'initialData', 'sportsApi'];
 
 	/* @ngInject */
-	function LeaguesController($modal, initialData, sportsApi) {
+	function LeaguesController(dialogs, initialData, sportsApi) {
 		/* jshint validthis: true */
 		var vm = this;
 
@@ -49,29 +49,12 @@
 
 		function deleteItem(id) {
 
-			var modalInstance = $modal.open({
-				templateUrl: 'app/shared/confirm-modal.html',
-				controller: 'confirmModalController',
-				controllerAs: 'vm',
-				resolve: {
-					data: function () {
-						return {
-							title: 'Delete?',
-							message: 'Are you sure you want to Delete?',
-							buttons: ['OK', 'Cancel']
-						};
-					}
-				},
-				size: 'sm'
-			});
-
-			modalInstance.result.then(function () {
-				sportsApi.deleteLeague(id).then(function(data){
-					_.remove(vm.leagues, { 'id': id });
+			dialogs.confirm('Delete?', 'Are you sure you want to Delete?', ['OK', 'Cancel'])
+				.then(function () {
+					sportsApi.deleteLeague(id).then(function (data) {
+						_.remove(vm.leagues, {'id': id});
+					});
 				});
-			}, function () {
-				console.log('Modal Dismissed');
-			});
 		}
 
 		function editItem(item) {
